userWithAddress endpoint construction, refined UserServiceImpl and UserRespository with JPQL query for users with addresses

+++ src/main/java/com/remolo/pizzeriadonremolo/repository/OrdersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.remolo.pizzeriadonremolo.entities.Orders;
 
-public interface OrdersRepository extends JpaRepository <Orders, Integer> {
+public interface OrdersRepository extends JpaRepository <Orders, Long> {
